Store file content on DidOpen
This allows Formatting to work on new files. The key is to not trigger
any error when the file path doesn't exists, even when parsing the file.

Errors from parsing are now stored in the ParsedGnoFile struct, the idea
is to eventually rely on it to publish diagnostics, instead of running
`gno precompile` manually, since the errors returned by `gno precompile`
also come from go/parser.ParseFile.

Fix #5